feat: Complete 03-FavouriteMovie
